Refactor existing components
